Moving assistors to their own folder
